Constructor mocking main implementation. Includes refactoring of JVM agent, but lacks Android refactoring

+++ mockk/src/main/kotlin/io/mockk/impl/instantiation/JvmObjectMockFactory.kt
@@ -10,43 +10,66 @@ import io.mockk.impl.stub.SpyKStub
 import io.mockk.impl.stub.StubGatewayAccess
 import io.mockk.impl.stub.StubRepository
 
-class JvmObjectMockFactory(val proxyMaker: MockKProxyMaker,
-                           val stubRepository: StubRepository,
-                           val gatewayAccess: StubGatewayAccess
+class JvmObjectMockFactory(
+    val proxyMaker: MockKProxyMaker,
+    val stubRepository: StubRepository,
+    val gatewayAccess: StubGatewayAccess
 ) : ObjectMockFactory {
+    val refCntMap = RefCounterMap<Any>()
 
-    override fun objectMockk(obj: Any, recordPrivateCalls: Boolean) {
-        val cls = obj::class
+    override fun objectMockk(obj: Any, recordPrivateCalls: Boolean): () -> Unit {
+        if (refCntMap.incrementRefCnt(obj)) {
+            val cls = obj::class
 
-        JvmStaticMockFactory.log.debug { "Creating object mockk for ${cls.toStr()}" }
+            JvmStaticMockFactory.log.debug { "Creating object mockk for ${cls.toStr()}" }
 
-        val stub = SpyKStub(cls, "object " + cls.simpleName, gatewayAccess, recordPrivateCalls)
+            val stub = SpyKStub(cls, "object " + cls.simpleName, gatewayAccess, recordPrivateCalls)
 
-        JvmStaticMockFactory.log.trace { "Building object proxy for ${cls.toStr()} hashcode=${InternalPlatform.hkd(cls)}" }
-        try {
-            proxyMaker.proxy(
-                cls.java,
-                emptyArray(),
-                JvmMockFactoryHelper.mockHandler(stub),
-                false,
-                obj
-            )
-        } catch (ex: MockKAgentException) {
-            throw MockKException("Failed to build object proxy", ex)
+            JvmStaticMockFactory.log.trace {
+                "Building object proxy for ${cls.toStr()} hashcode=${InternalPlatform.hkd(
+                    cls
+                )}"
+            }
+            val cancellable = try {
+                proxyMaker.proxy(
+                    cls.java,
+                    emptyArray(),
+                    JvmMockFactoryHelper.mockHandler(stub),
+                    false,
+                    obj
+                )
+            } catch (ex: MockKAgentException) {
+                throw MockKException("Failed to build object proxy", ex)
+            }
+
+            stub.hashCodeStr = InternalPlatform.hkd(cls.java)
+
+            stub.disposeRoutine = cancellable::cancel
+
+            stubRepository.add(obj, stub)
         }
 
-        stub.hashCodeStr = InternalPlatform.hkd(cls.java)
-
-        stubRepository.add(obj, stub)
+        return {
+            if (refCntMap.decrementRefCnt(obj)) {
+                val stub = stubRepository.remove(obj)
+                stub?.let {
+                    JvmStaticMockFactory.log.debug {
+                        "Disposing object mockk for ${obj::class.toStr()} hashcode=${InternalPlatform.hkd(
+                            obj
+                        )}"
+                    }
+                    it.dispose()
+                }
+            }
+        }
     }
 
-    override fun objectUnMockk(obj: Any) {
-        proxyMaker.unproxy(obj)
-
-        stubRepository.remove(obj)
-    }
-
-    override fun clear(obj: Any, answers: Boolean, recordedCalls: Boolean, childMocks: Boolean) {
-        TODO("clear")
+    override fun clear(
+        obj: Any,
+        answers: Boolean,
+        recordedCalls: Boolean,
+        childMocks: Boolean
+    ) {
+        stubRepository[obj]?.clear(answers, recordedCalls, childMocks)
     }
 }
